LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/7.9.2_Examples/S7.9.2_A1_T3-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected token ')'. Expected a ';' after the for loop condition expression.
S7.9.2_A1_T3

PASS Expected parsing failure

TEST COMPLETE
